app/backend/pyproject.toml
[project]
name = "app-security-audit-backend"
version = "0.1.0"
description = "Backend FastAPI para avaliacao read-only de seguranca e performance Linux"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "httpx>=0.28.0,<1.0.0",
  "psutil>=6.1.0,<7.0.0",
  "reportlab>=4.2.0,<5.0.0",
   "docker>=7.0.0,<8.0.0",
   "python-nmap>=0.7.1,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-cov>=6.0.0,<7.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
